fix: resolve all CI failures, security vulnerabilities, and import errors
- Updated starlette to 0.49.1 and fastapi to 0.136.1 to resolve CVEs.
- Created src/analytics/__init__.py and src/monitoring/__init__.py.
- Updated src/__init__.py and src/trading/__init__.py with full exports.
- Fixed ruff violations (RUF022, RUF046) and applied formatting.
- Resolved test failures in health, data pipeline, and reporting suites.
- Verified zero failures across linting, security, and test jobs.

+++ src/research/hyperopt_walkforward.py
@@ -237,17 +237,13 @@ class WalkForwardOptimizer:
 
                 for direction in [-1, 1]:
                     perturbed_params = params.copy()
-                    delta = (
-                        max(1, int(abs(original_val) * 0.05))
-                        if is_int
-                        else original_val * 0.05
-                    )
+                    delta = max(1, int(abs(original_val) * 0.05)) if is_int else original_val * 0.05
 
                     new_val = original_val + (direction * delta)
 
                     # Ensure type safety
                     if is_int:
-                        new_val = int(round(new_val))
+                        new_val = round(new_val)
 
                     # If no change occurred (e.g. value was small and delta was 0), skip
                     if new_val == original_val:
